Keep changed CWD with 'autochdir' setting by temporarily disabling it.
For example, I have a ":lcd .." in after/ftplugin/gitcommit.vim, so that I'm in the Git root directory, not the .git subdir when composing the commit message. The reload of the buffer by AutoFenc (via :edit) again triggered the automatic change of the working dir, and therefore my customization was lost.
The 'autochdir' setting needs to be temporarily disabled to avoid that.